Remove platform-specific logic for qtactiveqt module.
For the sake of a uniform test procedure, it is cleaner to have
`make module-qtactiveqt' always working, but simply not building
anything on platforms where ActiveQt is not supported.

Without this, any test procedure for building Qt5 needs to be able
to know or determine that ActiveQt should only be built on Windows.

Reviewed-by: Jason McDonald
Change-Id: I3869bbacdc6598a764561a4b9f6042c49e6ffae8