Can now tell init-repository to only use a subset of modules
Still useful to use init-repository because of all the other logic, such
as repository urls. More general than --no-webkit, although it is still
a valid convenience as the most common case.

Also makes qt.pro check for existence of modules, since it is no longer
certain that you'll have everything but webkit.

Change-Id: Iaaaa30e24e0a793218e2e307bdde56067760c513
Reviewed-on: http://codereview.qt.nokia.com/1258
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>