Prepare init-repository script for moving v8 submodule
Transitional step needed while v8 is moved from qtdeclarative
to qtbase.
Before the transition is completely in place, init-repository
must be able to cope with both qtbase and qtdeclarative having
a v8 submodule.
This change causes git submodule init/update to be called for
both repositories, even though one of them doesn't have any
submodules. But in that case it's a no-op.

Change-Id: Ia1af01308c808e62eb85431764edd8c12bdc3325
Reviewed-on: http://codereview.qt.nokia.com/3303
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>