Clean up build script
Remove loading of thread modules. Not core module, and we don't use
them.

Return value of system command.

Add --force-qmake option, and by default avoid running qmake when a
Makefile already exists.

Do not pass '-j 1' unless explicit on the command line, and pick up
'MAKE' environment variable. (Thiago)

Remove usage of File::Which, as it's not a core module, and often not
present on host machines. (Peppe)

Remove '-s' or '/s' optin to make/nmake, as we cannot see how far we
have compiled, and we only get warnings/errors.

Change the qtwebkit build command, and add separate command for
installing qtwebkit. Also add the required build tools for WebKit
on Windows.

Change-Id: I79bffa39a13ece78fa401f39a38a1ccaf0f389b0