Fix building on Windows and OSX
A problem with the dependency algorithm would make the top-most
dependent module not compile. Also affects building QtWayland on
Linux.

OSX has a problem when doing 'make install' on Qt configured with
prefix pointing to qtbase (no installation needed), so disable that
for the alpha, like on Windows.

Also, remove a redundant line in the script.

Change-Id: I368e80520977cf8a4bfcbbf13445f50c820aee9c
Reviewed-by: Lars Knoll <lars.knoll@nokia.com>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>