Follow the protocol requests for webkit (special treatment for internal)
It is currently not possible to successfully run the init-repository script with
certain protocols, like "http", since those requests does not apply for
qtwebkit. Hence, the cloning of qtwebkit fails behind the corporate rules which
is against the original purpose of those command line options.

If the requested protocol is "internal", then the replacement of qtwebkit
should still be excluded since that would potentially cause breakages.

Change-Id: I738738f1c367cf1b4f18ffed05f37715a78e94ae
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>