Remove qtjsondb from qt5.git
The jsondb module has been ongoing research work. But it won't be done
in time for Qt 5.0, and is currently also less relevant. It's also
currently not being actively worked on. So let's remove it
from the list of modules compiled with qt5.git and keep it as
research.

The module can still be compiled stand-alone on top of the Qt essentials.

Change-Id: I6fd4524cae43b803a842788ff22a7525dc7bc0e6
Reviewed-by: Yuchen Deng <loaden@gmail.com>
Reviewed-by: Kevin Simons <kevin.b.simons@gmail.com>