Fix make install with WebKit
Avoid running "make install" through build-webkit and therefore risk
accidentally rebuilding the lot. We call build-webkit either with --release
or without, but we never pass --debug. Therefore the build directory is
always WebKitBuild/Release.

This is a stop-gap for the Qt 5 beta. After the release we'll get rid of
qtwebkit.pri altogether and WebKit will be built with just qmake and make.

Change-Id: I21ff7ef8420b67cf466d5bca5efd7a1245bf4131
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>
Reviewed-by: Lars Knoll <lars.knoll@nokia.com>