Add win_flex to the convenience GNU tools for the WebKit build on Windows
The latest builds of ANGLE in WebKit require a version of flex that is newer
than what GnuWin32 can provide. The only distribution I've found that doesn't
pull in all of CygWin is winflexbison.

This patch adds the binary of it along with a README explaining the purpose
of this gnuwin32 directory altogether.

Change-Id: Ia00fa5d628891c9d4c30586f42274c7c602362ce
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>
Reviewed-by: Lars Knoll <lars.knoll@nokia.com>