WebKit update
This should fix the build errors that "looked" like errors from Qt3D
but were triggered through an incorrect QtWebKit module master header file,
as well as the supposed linking errors related to WebKit2.

Also disable the WebKit build if ICU is not available instead of breaking
the build when reaching WebKit.

Change-Id: I109b4152c081250c8950ef62dd7278129b716f76
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>