Fix Coin provisioning in RHEL 6 regarding sslclientkey values
There was a bug in RHELs subscription manager that caused yum
not to work by fetching incorrect data into the sslclientkey repository
parameter value.
This provisioning script fetches new data for the repository and thus
fixes yum. Naturally by not having a template where a bad repository
exists, this script would be obsolete. Also, as this needs to be run
before any other scripts using yum, prefixes were added to the scripts.

Change-Id: I5c6ed1d43451d2918f9b6dad22b0106339f36aea
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>