automatically launch new configure system in top-level builds
this includes moving the -skip option here, where it belongs.
the implementation is a bit hacky, but the proper one would be vastly
more complex: we'd have to load all modules, then parse the command
line properly (the -skip parsing would be a custom callback, to avoid
needing to process features), then discard all now excluded modules
again, then re-process the command line (so we can complain about
options which are now invalid), and only now proceed with the actual
feature processing.
that seems a bit over the top, so we accept that -skip cannot be the
argument of another option, which seems quite reasonable.

Change-Id: I0053658e6d7e11b82cdfe55893cb553b664a9298
Reviewed-by: Jake Petroules <jake.petroules@qt.io>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>