Clean up usage of qtsvg as dependency
There is no build dependency on QtSvg in qtdeclarative or qtquick1, so
don't recommend it for builds. QtSensors on the other hand has an
example that does need svg.
This does not change anything in the big picture, the svg module will be
built as always and svg files will be displayed. The change is the
ordering of builds (leave more room for parallelization).

Change-Id: I50e527e03a75742036d5f914611d0b34437a8e3a
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>