Provision: Set bash_profile to read .bashrc in macOS 10.12
The method (vmware tools) of launching a process inside a virtual
machine results in the launch of a process that does not originate
from a login shell. Therefore the CI system explicitly starts the
agent with "bash -l -c" to enforce the invocation as login shell
and thus execute settings from /etc/profile, ~/.bash_profile, etc.

This brings the behavior of the system closer to what we see as
developers.

However the shell is a non-interactive shell, therefore ~/.bashrc
is not read by default.

On Linux systems such as RHEL, ~/.bash_profile is configured to still
read ~/.bashrc.

For consistency we apply the same logic to the macOS configurations.

Task-number: QTQAINFRA-1091
Change-Id: I72206dbd8f7289de86830db23fa1ae2c548181eb
Reviewed-by: Heikki Halmet <heikki.halmet@qt.io>