Add preparing of qemu fontconfig to Ubuntu 16.04 provisioning
If normal fontconfig paths are used, qemu parses what ever files it finds from
the toolchain sysroot and the rest from the system fonts. Fix by copying the
system font configurations to a location which prefix that can't be found from
the toolchain sysroot. Links must also be dereferenced or their targets remain
pointing to the toolchain sysroot.

Task-number: QTBUG-59966
Change-Id: I9710c9fdebe81114080f47f1f59b2a049f87c070
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@qt.io>