Turn off line ending conversion for .commit-template
When committing a change, the automatically generated git commit message
consists of

1) The content of .commit-template
2) Informative text for the current commit, generated by git.

While 2) always has LF line endings, even on Windows, 1) might have CRLF
line endings if core.autocrlf=true. The result is a text with mixed line
endings. With the default editor vim, every line of the commit template
part ends with ^M.

Fix this by forcing the line endings of .commit-template to be LF
regardless of core.autocrlf.

Change-Id: I6db825695466d7f48f1fa4f771d83fb872fdb576
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>