Enable developer mode on macOS platforms
Previously when a VM user tried to attach lldb to a process via ssh,
lldb would fail with the following error message:
"error: process exited with status -1 (unable to attach)". If this
were done over VNC rather than ssh, the user would instead get a UI
popup to enter a password to allow debugging the process.

When trying to debug an application using Xcode for the first time,
Xcode will offer to enable the so called "Developer Mode" which allows
attaching a debugger without specifying a password every time.

This change enables "Developer Mode", so that every VM user doesn't
have to.

Change-Id: Ie4e473a982c908afd456ac1773bbdaf4ce778094
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>