provisioning: install MinGW 530 and 630 for Win10 x64
MinGW is an alternative option for Visual Studio. It's
used to create desktop builds and for cross compiling
Android and QNX. At this point it is still unclear which
version is going to be used for what, but as both are
supported, let's think ahead and provide them both.
This also allows us to debug with the other, if the
other is causing problems.

Task-number: QTQAINFRA-1289
Change-Id: I2a922669c07d49370e608c093be80481c8613dc7
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>