Fix Ubuntu 16.04 random provisioning fail
Ubuntu 16.04 provisioning sometimes fails in the apt-get commands for
package installations. This is caused by a lock that is still in use
when the system is calling the apt-get commands.

This change will use the apt daemon to queue the install commands
and wait for the lock to release instead of checking once and then
failing if the lock is not open.

Task-number: QTQAINFRA-1502
Change-Id: Idd4aff42f99cf4d14c08aef8fcc3cb7dbc58b2dd
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@qt.io>