Simplify apt package installation and disable daily package maintenance
Instead of using aptdcon, we can continue to use apt-get to install the
software. We just need to make sure that the automatic update and
package upgrade services are disabled. This is generally desirable as we
do not want these kind of activities to occur in the background while
building or running tests.

This also fixes the broken package installation via regex, which apt-get
supports but aptdcon does not.

Revert "Fix Ubuntu 16.04 random provisioning fail"

This reverts commit 537ea17d14526bc6564765e2dd4652e11a3ed235.

 Conflicts:
	coin/provisioning/qtci-linux-Ubuntu-16.04-x86_64/002-apt.sh

Change-Id: I920e5132bb0a248d3794e1efcf01ce78d28e6ee3
Reviewed-by: Joni Jäntti <joni.jantti@qt.io>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@qt.io>