Fix opcua provisioning by sourcing .profile
When CMake is installed it exports its install
path into .profile. However, freeopcua doesn't source
it as builds do, so it couldn't find the cmake.
Also, as Ubuntu is using .profile in our builds,
this commit fixes so that the PATH is actually written
into the right file.

Task-number: QTBUG-65594
Change-Id: I561a88a07872436bb4a3bd62a2d55fd56a130121
Reviewed-by: Joni Jäntti <joni.jantti@qt.io>
Reviewed-by: Heikki Halmet <heikki.halmet@qt.io>