Use "correct" registry when enabling Windows 10 Developer Mode
As coin is running a 32 bit application the registry key ended
up in the Wow6432Node compatibility area of the registry when run
on a 64 bit Windows. Windows does not use this place to check
for the developer mode setting, so we use the "proper" registry
instead.

If adding the registry key fails the script now reflects that
and exits with a proper exit code.

Change-Id: Ifd6bc7e7b0ed6595a2a967597ab45053cb51e7aa
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>