Provisioning: Add test for squish installation
Squish is needed by RTA which is not run during normal qt5 build. Other
words squish installation is not tested during integration. That's why
we need separate test for it.
This change will fetch test and run test suite which is made to check
that squish installation works

Task-number: QTQAINFRA-1733
Change-Id: I9284fd52d87767ae04022c3a6996e0bb972315dd
Reviewed-by: Johanna Äijälä <johanna.aijala@qt.io>
Reviewed-by: Joni Jäntti <joni.jantti@qt.io>
Reviewed-by: Matti Paaso <matti.paaso@qt.io>