Install libusbx-devel for QDB on RHEL 7.4
On previous RHEL version the package was libusb1-devel. That package doesn't
seem to be there anymore, which presumably means why the package to install was
changed to libusb-devel. Unfortunately libusb-devel provides a compatibility
package for the old 0.1 API version and doesn't work with QDB. libusbx-devel
seems to be the new name for the 1.0 API version of libusb on RHEL.

(cherry picked from commit 140c6479b5916e21d6c03a94f844a0f8693af0f8)
Change-Id: I978e9842b3f0ec3a355cd1724934c99a531903ef
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>