Provisioning: Ubuntu - remove Unity window decorations
Unity does not respect setGeometry() call, which is
violation of WM spec. During the years we have tried
various workarounds, but it has proved to be counterproductive.

This patch adjusts Unity theme to not draw window decorations,
ensuring that Unity treats the setGeometry() call the same way
all the other WMs and platforms we are testing on.

For more detailed explanation see QTBUG-67757.

Task-number: QTBUG-67757
Change-Id: I8d2bb9f2af14d3c184fc218ed08a1febd70f4026
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>