Install the MinGW build of libclang
This restores the MinGW installation of libclang on Windows 7 and
installs the MSVC and MinGW builds in parallel on Windows 10. For
compatibility reasons the MSVC installation is done last, to ensure that
the older LLVM_INSTALL_DIR variable points to the MSVC installation. In
addition though we now get LLVM_INSTALL_DIR_MINGW and
LLVM_INSTALL_DIR_MSVC, which qt_find_clang.prf in qttools can use in the
future.

Task-number: QTBUG-66015
Change-Id: I8b53b85d0cde597c88a089f2a05f03947828d9bd
Reviewed-by: Simo Fält <simo.falt@qt.io>