Update winrtrunner
A newer build of winrtrunner is needed. It is built from dev and
contains fixes related to handling of test output (most
importantly 044e9ad878ed1cf0ace0a9260b55cc94c71a294d).

To make the script more stable, the newer version of winrtrunner
was also copied to ci-files01, so that an alternative download
location can be offered.

Change-Id: I6a429aa3292266067756ce7ba3553088dc168197
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>