Docker Provisioning: Install Avahi and enable mDNS service discovery
The challenge of connecting Docker containers from the host is to
resolve a predefined hostname to an IP address assigned at runtime. For
Unix-like platforms such as Ubuntu, it can be done by either hard-code
the IP addresses or creating a DNS server. Both of them are not
user-friendly which need root permission to modify system files.

Avahi is a free Zeroconf implementation and provides multicast DNS
(mDNS) service discovery. It enables Docker containers to publish their
services on a local network. For the host systems supporting mDNS
technology, they will automatically discover the services in a local
domain without extra user-efforts.

Task-number: QTQAINFRA-1686
Change-Id: I1ef464a455cef5d331a3e2eb0194e2741172ef95
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Ryan Chu <ryan.chu@qt.io>
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@qt.io>