fix setting QTDIR_build for default_pre.prf
adding it to the super cache makes no sense, as that's read before
spec_pre.prf, which clears CONFIG. use the regular cache instead.

submodules do not inherit that setting, and they don't need to as per
comment.

also, don't keep adding it to the cache - the cache() function doesn't
check whether the added value is already present, so we need to do that
manually.

amends 31590c0b65.

Fixes: QTBUG-63452
Change-Id: I608559a80a7c772dbdaacfaaadb4fc268b442c39
Reviewed-by: Kai Koehne <kai.koehne@qt.io>