Docker Provisioning: Remove the fixed version of package dependencies
Free the dependencies of the specific Ubuntu packages. Ensure that test
server is using the latest version of the Ubuntu packages to test
network changes. Otherwise, all the docker files need to be manually
updated when upgrading to the next Ubuntu version (e.g. 18.04).

For debugging purpose, the installed packaged will be listed in the
provisioning logs.

Change-Id: I38d119d11f33cfd60dc34fcd57ec0fe1615e98f2
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>