provisioning: consistently install libxkbcommon* on linux distros
Ubuntu was the only distro that installed both of the required packages:

installPackages+=(libxkbcommon-dev)
installPackages+=(libxkbcommon-x11-dev)

RHEL 7.4 did not explicitly install any, but got libxkbcommon-dev
pulled in as gtk3 dependency. Extract from provisioning log:

--> Processing Dependency: pkgconfig(xkbcommon) >= 0.2.0 for package: gtk3-devel-3.22.30-3.el7.x86_64
---> Package libxkbcommon-devel.x86_64 0:0.7.1-1.el7 will be installed

openSUSE explicitly installed libxkbcommon-devel, but not libxkbcommon-x11-devel.

Qt official binaries are built on RHEL, where we do not have
libxkbcommon-x11-dev installed. This means that XCB plugin and
Compose input plugin would fallback to using bundled sources.
It was actually desired until now that XCB/Compose use the
bundled sources instead of linking with the library from the
system, but this was only a lucky side effect. If we had installed
libxkbcommon-x11-dev on RHEL, then Qt binaries would end up
linking with the libxkbcommon from the system, because of the
missing "-qt-xkbcommon" configre switch for release builds.
We won't bundle libxkbcommon anymore, hence we need to install
the missing dependencies, as done by this patch.

Change-Id: I5c7b8ac38c266ce81cb5a3189a9082bfd581ee31
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@qt.io>