Prospective fix for git lfs installation
Keys used to sign package repositories tend to have expiry dates and
therefore need to be changed from time to time. Therefore hardcoding
the key we expect here will break (as it does right now) and it's better
to retrieve the expected key via a secure transport.

The old key - per
http://keyserver.ubuntu.com/pks/lookup?search=0xC2E73424D59097AB&op=vindex
- will expire on the 12 of January. That's probably why they changed the
key used to sign the repo before the expiry.

Change-Id: I39d082a6e12731b7eb0a8214a500101ca24f8af4
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>
Reviewed-by: Heikki Halmet <heikki.halmet@qt.io>