Update qtsaferenderer platform configuration definitions
The project-specific platform configuration file is renamed to follow the
convention as for other projects. This will allow Coin to recognize the
configuration that are specific for qtsaferenderer project where they
were previously (unintentionally) ignored.

Removed QNX 7 as it is in the Qt mainline configuration.

Removed Ubuntu 16.04 as Ubuntu 18.04 is now supported.

Task-number: QSR-209
Change-Id: I505edc3741dbc047b2fc23bc409016d712596872
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@qt.io>