Remove platforms from Qt 6
Get to a smaller set to reduce load on the CI and have faster
turnaround times while developing. Remove some older compilers
that do not support the minimum C++ feature set we want to require
for Qt 6.

We'll need to revisit and probably expand this set when we get closer
to releasing Qt 6.

Change-Id: I6afba8b3caca0fac5bbf80efdcde7714b8d1ab76
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>