Update vcpkg installation
Use the latest snapshot, fetch the package set from the repo itself and
set the VCPKG_ROOT environment variable as that's the vcpkg documented
interface.

The qt specific triplet defaults to static linkage, which allows
removing the %PATH% handling and resembles the same setup as in Qt 5
where we linked the 3rd-party code statically into qt.

Change-Id: I789d61b5c7258674c2dd80403efaeb891c8f75c7
Reviewed-by: Liang Qi <liang.qi@qt.io>