platform_configs: add -bundled-xcb-xinput to RHEL-7.6:Packaging
All libs, expect xcb-xinput have been dropped from 3rd party sources
in qtbase. When building default binaries for releases we used to have
"qt-xcb" switch to reduce dynamic depenendies of the binary. We still
need to do that for xcb-xinput (via -bundled-xcb-xinput) as this library
is not available on some of the supported distributions.

Change-Id: Iece12bc38978ddb66f84928f661924f076c3647e
Reviewed-by: Liang Qi <liang.qi@qt.io>