Work around cmake issue breaking C++ feature detection on Android
As per https://gitlab.kitware.com/cmake/cmake/issues/19515, the version
of CMake we're currently using in the CI has an issue detecting clang as
GNU compatible C++ frontend and thus fails to convert the values in the
CMAKE_C/CXX_STANDARD variables to compiler flags correctly. Work around
this for now as suggested in the upstream ticket. We can remove this
when upgrading to a newer cmake version.

Change-Id: I0ea37503c086ff86e05c50b812474522a7737b20
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>