Change local Ubuntu repo mirror in provisioning
The old mirror host Qt company used to sync their local mirror
from, stopped syncing from Ubuntu in October 2018.
The mirror mechanism was changed so that in the future if
addresses change, we don't have to touch the provisioning scripts
as the address is more general. Also a duplicate entry was removed.

Task-number: QTQAINFRA-3279
Change-Id: I65ada052d308eaef0411eb9caad893304eb05ab6
Reviewed-by: Heikki Halmet <heikki.halmet@qt.io>