Reduce the set of modules for Qt6 development
Change the amount of modules that are tested together by the CI (and
then used for initial Qt 6 snapshots) to the core essential modules. All
others are set to "ignore" and the objective is to bring them back later
when qtbase/declarative/quick3d development has settled.

Notable exceptions are svg and graphicaleffects due to their strong
dependency from qtvirtualkeyboard as well as qtquickcontrols2.

Change-Id: I972d7b353add1f5dbad5104ce84c4cc5b8c6f616
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>