Add '-qt-zlib' as a configure argument for windows builds
This is needed to use the bundled zlib in windows builds. In Qt 5.14.0
windows packages were using bundled one but in Qt 5.14.1 not anymore. So
this change is to restore same behavior than in Qt 5.14.0.

Linux and macOS packages seems to use system zlib also in 5.14.0; that's
why no changes needed there

Task-number: QTBUG-81724
Change-Id: Ife531d96c8438491e5d7b1f97f5338c4627be7e8
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>