Mask PackageKit on SUSE
Fixes a flaky conflict between zypper and packagekitd.

packagekit is a *static* unit, that means it is missing an [Install] section,
so it can't be enabled or disabled. It is started as a dependency by other
units, or by dbus activation.

We *mask* the unit to make it impossible to start in any way.

For that we had replace a few occurrences of "pkcon" with the native package
manager.

Finally the package "libpython3_4m1_0" is not being installed any longer
because it never was: this package does not exist. It's just that pkcon does
not fail on invalid package names.

Fixes: QTQAINFRA-3523
Change-Id: I6ff580afb05aa508042c444b2ce777c3a9443dfe
Reviewed-by: Heikki Halmet <heikki.halmet@qt.io>