Transition to macOS 10.14 step 1
We want to run all tests on macOS 10.14, which requires some more fixes
in qtquickcontrols2. That however isn't passing due to unrelated changes
in base/declarative blocking us from moving forward. So to decouple the
two, this change limits macOS tests to qtbase and declarative but on
10.14. This will allow us to switch qtbase over to C++17 and once we've
got qqc2 updated in qt5.git, we can globally re-enable tests for 10.14.

Task-number: QTQAINFRA-3108
Change-Id: I0f835a1da87187c6bb69e3d4019196e1411095fd
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>