CMake: Change iOS builds to build in Release configuration
Apparently standalone Debug iOS builds are not supported in qmake
at the moment as per QTBUG-71990.

While we keep the same configure.json error reports for both qmake and CMake
it's simpler to switch the config to Release.
Once we get Ninja Multi-Config in Coin working, we can switch to
debug_and_release.

Change-Id: I406012918fd402ad0ff29bbdc151e247699eea72
Reviewed-by: Cristian Adam <cristian.adam@qt.io>