CMake: Enable qt5.git CMake top-level / super builds
Add simple instructions to only configure and not build top-level
builds, similar to what the qmake configurations do.

The CMake configurations need to be copy-pasted again into qt5.yaml.
We can't use default.yaml because that would enable CMake
configurations for repositories that are not ported to CMake, and
thus break the integrations.

Update qtqa sha1 to a dev branch one which contains the
module_config.yaml instructions to build qtqa with CMake.

Task-number: QTBUG-84874
Task-number: QTBUG-84918
Change-Id: I8c64f5e0acc96e14a4952abb6c77fe5c8e27a0ec
Reviewed-by: Liang Qi <liang.qi@qt.io>