CMake: Change armv7 configuration to prepare for building tests
Switch the configuration to have release + developer-build similar
to the qmake configurations.

Add debug info for extra spiciness.

Specify a new env variable called QT_SYSROOT_FOR_RUNNING_TESTS which
will be used by the qtbase instructions to set QEMU_LD_PREFIX, which
is needed to run the tests via qemu.

Specify the QT_QUICK_BACKEND=software env var as well as the
GST_PLUGIN_SCANNER env var, because the qmake configurations do it.

Make sure to cross-build tools, because some of the tests like
tst_uic and tst_rcc expect the tools to be there.

Note that DisableTests feature is left on purpose, so we can decouple
the building from the testing in 2 different steps.

Task-number: QTBUG-84423
Task-number: QTBUG-83968
Change-Id: Icf6b12e6942c6ae6a0445eeac3f88704ca797148
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>