CMake: Switch qtbase Windows config to use Ninja Multi-Config
And configure with two configurations, RelWithDebInfo and Debug.
This is equivalent to qmake's
--debug-and-release --force-debug-info

This should give us minimal coverage for debug-and-release builds,
as well as avoid regressions with NMC aka multiple configurations.

If the qtbase builds successfully, we can try to extend the rest of
the repositories to also build with NMC on Windows.

We can't add such a configuration to macOS for now, because it
requires CMake 3.18, which is not released yet.

Disable PCH because it seems to fail the build when used together
with Ninja Multi-Config.

Task-number: QTBUG-80900
Change-Id: I71a9584ddb1d48ed22925f67b50891e102036c97
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>